[feat] Calculate Total Score
