+++ web/src/app/pages/welcome/welcome.component.html
@@ -30,7 +30,7 @@
   </div>
 
   <div
-    class="col-span-12 grid grid-cols-2 gap-x-3"
+    class="col-span-12 grid grid-cols-1 gap-x-3"
     *ngIf="(auth.user$ | async) !== null"
   >
     <a mat-button color="primary" routerLink="/app/">Dashboard</a>
