+++ web/src/app/pipes/arrays-equal/array-equal.module.ts
