+++ web/src/app/pipes/arrays-equal/array-equal.pipe.ts
