+++ web/src/app/pipes/arrays-equal/index.ts
