+++ web/src/app/utils/equals.ts
@@ -1,0 +1,5 @@
+export type Equals<T> = (a: T, b: T) => boolean;
+
+export function identityEquals<T>(a: T, b: T): boolean {
+  return a === b;
+}
