+++ web/src/app/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './rxjs';
+export * from './arraysEqual';
