+++ web/tsconfig.json
@@ -23,6 +23,7 @@
       "components/*": ["src/app/components/*"],
       "api/*": ["src/app/api/*"],
       "utils/*": ["src/app/utils/*"],
+      "pipes/*": ["src/app/pipes/*"],
       "environments/*": ["src/environments/*"]
     }
   },
